+++ eyes_core/applitools/core/eyes_mixins.py
@@ -1,10 +1,4 @@
-from typing import Optional, Text
-
 from applitools.common import Configuration, logger
-from applitools.core.debug import (
-    FileDebugScreenshotsProvider,
-    NullDebugScreenshotsProvider,
-)
 
 
 class EyesConfigurationMixin(object):