+++ eyes_images/applitools/images/eyes.py
@@ -1,6 +1,7 @@
 import typing
 
 from applitools.common import Configuration, EyesError, RectangleSize, Region, logger
+from applitools.common.utils.compat import basestring
 from applitools.common.utils.general_utils import all_fields, proxy_to
 from applitools.core import (
     NULL_REGION_PROVIDER,
@@ -73,20 +74,33 @@ class Eyes(EyesBase):
         # type: (Text, Text, Optional[ViewPort]) -> None
         self.open_base(app_name, test_name, dimension)
 
+    @typing.overload
     def check(self, name, check_settings):
         # type: (Text, ImagesCheckSettings) -> bool
+        pass
+
+    @typing.overload
+    def check(self, check_settings):
+        # type: (ImagesCheckSettings) -> None
+        pass
+
+    def check(self, check_settings, name=None):
         if self.configure.is_disabled:
             return False
+        if isinstance(name, ImagesCheckSettings) or isinstance(
+            check_settings, basestring
+        ):
+            check_settings, name = name, check_settings
         if name:
             check_settings = check_settings.with_name(name)
-        else:
-            name = check_settings.values.name
 
         image = check_settings.values.image
         if self.configure.viewport_size is None:
             self.configure.viewport_size = RectangleSize.from_(image)
 
-        return self._check_image(NULL_REGION_PROVIDER, name, False, check_settings)
+        return self._check_image(
+            NULL_REGION_PROVIDER, check_settings.values.name, False, check_settings
+        )
 
     def check_image(self, image, tag=None, ignore_mismatch=False):
         # type: (Union[Image.Image, Text], Optional[Text], bool) -> Optional[bool]